pyproject.toml
[tool.poetry]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.10.0,<3.12"
fastapi = "^0.112.0"
uvicorn = "0.23.2"
sqlmodel = "^0.0.24"
pydantic = "^2.11.7"
sqlalchemy = "^2.0.41"
passlib = "^1.7.4"
typing-extensions = "^4.14.1"
jose = "^1.0.0"
python-dotenv = "^1.1.1"

[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"